# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Bookstore API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/departemen-swagger": {
            "get": {
                "tags": [
                    "DepartemenSwagger"
                ],
                "summary": "Menampilkan semua departemen",
                "operationId": "01b3bd4684a21c78807dd561b9fdc36f",
                "responses": {
                    "200": {
                        "description": "Data departemen ditemukan"
                    },
                    "500": {
                        "description": "Kesalahan server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "DepartemenSwagger"
                ],
                "summary": "Membuat departemen baru",
                "operationId": "f2800d036c5925cebdae80de015959df",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_departemen",
                                    "kepala_departemen",
                                    "jumlah_karyawan",
                                    "keterangan"
                                ],
                                "properties": {
                                    "nama_departemen": {
                                        "type": "string"
                                    },
                                    "kepala_departemen": {
                                        "type": "string"
                                    },
                                    "jumlah_karyawan": {
                                        "type": "integer"
                                    },
                                    "keterangan": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Departemen berhasil dibuat"
                    },
                    "400": {
                        "description": "Permintaan tidak valid"
                    },
                    "500": {
                        "description": "Kesalahan server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/departemen-swagger/{id}": {
            "get": {
                "tags": [
                    "DepartemenSwagger"
                ],
                "summary": "Menampilkan detail departemen",
                "operationId": "36e7c6c5922edfa89f03ee247bb4118b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data departemen ditemukan"
                    },
                    "404": {
                        "description": "Departemen tidak ditemukan"
                    },
                    "500": {
                        "description": "Kesalahan server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "DepartemenSwagger"
                ],
                "summary": "Memperbarui departemen",
                "operationId": "8f4b98fff21e577a35f595fe1b1960c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_departemen": {
                                        "type": "string"
                                    },
                                    "kepala_departemen": {
                                        "type": "string"
                                    },
                                    "jumlah_karyawan": {
                                        "type": "integer"
                                    },
                                    "keterangan": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Departemen berhasil diperbarui"
                    },
                    "404": {
                        "description": "Departemen tidak ditemukan"
                    },
                    "400": {
                        "description": "Permintaan tidak valid"
                    },
                    "500": {
                        "description": "Kesalahan server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "DepartemenSwagger"
                ],
                "summary": "Menghapus departemen",
                "operationId": "17deb7dd49becb24b464e519d3164392",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Departemen berhasil dihapus"
                    },
                    "404": {
                        "description": "Departemen tidak ditemukan"
                    },
                    "500": {
                        "description": "Kesalahan server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/departemen-swagger/{id}/karyawan": {
            "get": {
                "tags": [
                    "DepartemenSwagger"
                ],
                "summary": "Menampilkan detail departemen beserta karyawannya",
                "operationId": "ecc232475822309392f48f81cf485053",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data ditemukan"
                    },
                    "404": {
                        "description": "Departemen tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication using Laravel Sanctum",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "security": [
                {
                    "sanctum": []
                }
            ]
        }
    },
...
}